if exists

+++ db/migrate/004_restart.rb
@@ -1,9 +1,9 @@
 class Restart < ActiveRecord::Migration
   def change
     # This essentially defeats the purpose of having a migration.
-    drop_table :commits
-    drop_table :entries
-    drop_table :repos
+    drop_table :commits unless !ActiveRecord::Base.connection.table_exists? 'commits'
+    drop_table :entries unless !ActiveRecord::Base.connection.table_exists? 'entries'
+    drop_table :repos unless !ActiveRecord::Base.connection.table_exists? 'repos'
 
     create_table :commits do |t|
       t.string :repo
